+++ .idea/misc.xml
@@ -1,7 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="EntryPointsManager">
     <list size="2">
-      <item index="0" class="java.lang.String" itemvalue="badgerlog.entry.Entry" />
+      <item index="0" class="java.lang.String" itemvalue="badgerlog.annotations.Entry" />
       <item index="1" class="java.lang.String" itemvalue="badgerlog.networktables.mappings.MappingType" />
     </list>
     <writeAnnotations>
