+++ lib/src/main/java/badgerlog/annotations/Entry.java
@@ -1,4 +1,4 @@
-package badgerlog.entry;
+package badgerlog.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
