+++ lib/src/main/java/badgerlog/processing/TestAspect.java
@@ -1,7 +1,6 @@
-package badgerlog;
+package badgerlog.processing;
 
 import org.aspectj.lang.JoinPoint;
-import org.aspectj.lang.ProceedingJoinPoint;
 import org.aspectj.lang.annotation.Aspect;
 import org.aspectj.lang.annotation.Before;
 import org.aspectj.lang.annotation.Pointcut;
